updated config and cache
