+++ src-tauri/src/cache/types.rs
@@ -1,3 +1,4 @@
+use std::fmt::Display;
 use serde::{Deserialize, Serialize};
 use ts_rs::TS;
 
@@ -5,4 +6,10 @@ use ts_rs::TS;
 #[ts(export, export_to = "../src/entity/cache/page.ts")]
 pub struct Page {
 	pub name: String,
+}
+
+impl Display for Page {
+	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+		write!(f, "{}", self.name)
+	}
 }
