+++ src-tauri/src/config/mod.rs
@@ -1,14 +1,15 @@
 use error_stack::{Result, ResultExt};
-
-use crate::config::config::ConfigError;
+use crate::config::error::ConfigError;
+use crate::config::main::Config;
 
 pub mod types;
-pub mod config;
+pub mod main;
+mod error;
 
-pub fn create() -> Result<config::Config, ConfigError> {
+pub fn create() -> Result<Config, ConfigError> {
 	let config_path = crate::dirs::create_conf_toml()
 			.attach_printable("error getting config json path")
 			.change_context(ConfigError)?;
 	
-	Ok(config::Config::create(config_path))
+	Ok(Config::create(config_path))
 }
